+++ src/main/java/com/onlineshop/modele/Panier.java
@@ -1,16 +1,15 @@
 package com.onlineshop.modele;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 public class Panier {
     private int id;
     private int clientId; // Référence au client
     private List<Produit> produits; // Liste des produits dans le panier
-    private BigDecimal total; // Total du panier
+    private double total; // Montant Total du panier
 
     // Constructeur
-    public Panier(int id, int clientId, List<Produit> produits, BigDecimal total) {
+    public Panier(int id, int clientId, List<Produit> produits, double total) {
         this.id = id;
         this.clientId = clientId;
         this.produits = produits;
@@ -42,11 +41,11 @@ public class Panier {
         this.produits = produits;
     }
 
-    public BigDecimal getTotal() {
+    public double getTotal() {
         return total;
     }
 
-    public void setTotal(BigDecimal total) {
+    public void setTotal(double total) {
         this.total = total;
     }
